fix(stream): preserve readable chunk boundaries
